assertion dtype bug fix

+++ dmc2gym/wrappers.py
@@ -144,7 +144,7 @@ class DMCWrapper(core.Env):
         self._observation_space.seed(seed)
 
     def step(self, action):
-        assert self._norm_action_space.contains(action)
+        assert self._norm_action_space.contains(action.astype(self._norm_action_space.dtype))
         action = self._convert_action(action)
         assert self._true_action_space.contains(action)
         reward = 0
